Fix bugs to make multi-datacenter work

+++ ryu/app/inception_rpc.py
@@ -14,7 +14,12 @@
 #    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 #    License for the specific language governing permissions and limitations
 #    under the License.
+from oslo.config import cfg
+
 from ryu.app import inception_conf as i_conf
+
+CONF = cfg.CONF
+CONF.import_opt('zookeeper_storage', 'ryu.app.inception_conf')
 
 
 class InceptionRpc(object):
@@ -24,8 +29,9 @@ class InceptionRpc(object):
         self.inception = inception
 
         # name shortcuts
-        self.mac_to_position = self.inception.mac_to_position
-        self.mac_to_ip = self.inception.mac_to_ip
+        self.arp_mapping = inception.arp_mapping
+        self.topology = inception.topology
+        self.flow_manager = inception.flow_manager
         self.dpid_to_conns = self.inception.dpid_to_conns
         self.dcenter_to_info = self.inception.dcenter_to_info
         self.vmac_to_queries = self.inception.vmac_to_queries
@@ -39,7 +45,7 @@ class InceptionRpc(object):
     def update_arp_mapping(self, ip, mac, dcenter):
         """Update remote ip_mac mapping"""
 
-        self.inception_arp.update_arp_mapping(ip, mac, dcenter)
+        self.inception.update_arp_mapping(ip, mac, dcenter)
 
     def send_arp_reply(self, src_ip, src_mac, dst_ip, dst_mac):
         self.inception_arp.send_arp_reply(src_ip, src_mac, dst_ip, dst_mac)
@@ -47,8 +53,11 @@ class InceptionRpc(object):
     def broadcast_arp_request(self, src_ip, src_mac, dst_ip, dpid):
         self.inception_arp.broadcast_arp_request(src_ip, src_mac, dst_ip, dpid)
 
-    def update_position(self, mac, dcenter, dpid, port, vmac):
-        self.inception.update_position(mac, dcenter, dpid, port, vmac)
+    def update_position(self, mac, dcenter, dpid, port):
+        self.inception.update_position(mac, dcenter, dpid, port)
+
+    def update_vmac(self, mac, vmac):
+        self.inception.vmac_manager.update_vm_vmac(mac, vmac)
 
     def del_tenant_filter(self, dpid, mac):
         self.inception.flow_manager.del_tenant_filter(dpid, mac)
@@ -61,36 +70,40 @@ class InceptionRpc(object):
         """
         # Failover logging
         log_tuple = (dpid_old, mac, vmac_old, vmac_new, dcenter_new)
-        self.inception.create_failover_log(i_conf.RPC_REDIRECT_FLOW, log_tuple)
+        if CONF.zookeeper_storage:
+            self.inception.create_failover_log(i_conf.RPC_REDIRECT_FLOW,
+                                               log_tuple)
 
-        if dcenter_new == self.inception.dcenter:
+        if dcenter_new == self.inception.dcenter_id:
             # This RPC is only used for multi-datacenter migration
-            self.inception.delete_failover_log(i_conf.RPC_REDIRECT_FLOW)
+            if CONF.zookeeper_storage:
+                self.inception.delete_failover_log(i_conf.RPC_REDIRECT_FLOW)
             return
 
         # Redirect local flow
-        gateway_dpid = self.inception.gateway
-        gateway_ip = self.inception.dpid_to_ip[gateway_dpid]
-        fwd_port = self.inception.dpid_to_conns[dpid_old][gateway_ip]
+        gateway_dpid = self.topology.gateway
+        fwd_port = self.topology.dpid_to_dpid[dpid_old][gateway_dpid]
 
-        self.inception.set_local_flow(dpid_old, vmac_old, vmac_new, fwd_port,
-                                      False)
+        self.flow_manager.set_local_flow(dpid_old, vmac_old, vmac_new,
+                                         fwd_port, False)
 
         if vmac_old not in self.vmac_to_queries:
             # The vmac_old key does not exist, indicating a duplicate RPC
-            self.inception.delete_failover_log(i_conf.RPC_REDIRECT_FLOW)
+            if CONF.zookeeper_storage:
+                self.inception.delete_failover_log(i_conf.RPC_REDIRECT_FLOW)
             return
 
         # send gratuitous ARP to all local sending guests
         # TODO(chen): Only within ARP entry timeout
         for mac_query in self.vmac_to_queries[vmac_old]:
-            ip = self.mac_to_ip[mac]
-            ip_query = self.mac_to_ip[mac_query]
+            ip = self.arp_mapping.get_ip(mac)
+            ip_query = self.arp_mapping.get_ip(mac_query)
             self.inception_arp.send_arp_reply(ip, vmac_new, ip_query,
                                               mac_query)
         del self.vmac_to_queries[vmac_old]
 
-        self.inception.delete_failover_log(i_conf.RPC_REDIRECT_FLOW)
+        if CONF.zookeeper_storage:
+            self.inception.delete_failover_log(i_conf.RPC_REDIRECT_FLOW)
 
     def set_gateway_flow(self, mac, vmac_old, vmac_new, dcenter_new):
         """Update gateway flow to rewrite an out-of-date vmac_old and
@@ -98,28 +111,32 @@ class InceptionRpc(object):
 
         # Failover logging
         log_tuple = (mac, vmac_old, vmac_new, dcenter_new)
-        self.inception.create_failover_log(i_conf.RPC_GATEWAY_FLOW, log_tuple)
+        if CONF.zookeeper_storage:
+            self.inception.create_failover_log(i_conf.RPC_GATEWAY_FLOW,
+                                               log_tuple)
 
         if vmac_old not in self.vmac_to_queries:
             # The vmac_old key does not exist, indicating a duplicate RPC
-            self.inception.delete_failover_log(i_conf.RPC_GATEWAY_FLOW)
+            if CONF.zookeeper_storage:
+                self.inception.delete_failover_log(i_conf.RPC_GATEWAY_FLOW)
             return
 
-        if dcenter_new == self.inception.dcenter:
+        if dcenter_new == self.inception.dcenter_id:
             # This RPC is only used for multi-datacenter migration
-            self.inception.delete_failover_log(i_conf.RPC_GATEWAY_FLOW)
+            if CONF.zookeeper_storage:
+                self.inception.delete_failover_log(i_conf.RPC_GATEWAY_FLOW)
             return
 
-        gateway = self.inception.gateway
-        _, remote_gw_ip = self.dcenter_to_info[dcenter_new]
-        fwd_port = self.dpid_to_conns[gateway][remote_gw_ip]
-        self.inception.set_local_flow(gateway, vmac_old, vmac_new, fwd_port)
+        gateway = self.topology.gateway
+        fwd_port = self.topology.gateway_to_dcenters[gateway][dcenter_new]
+        self.flow_manager.set_local_flow(gateway, vmac_old, vmac_new, fwd_port)
 
         # Send gratuitous arp to all guests that have done ARP requests to mac
         for mac_query in self.vmac_to_queries[vmac_old]:
-            ip_query = self.mac_to_ip[mac_query]
-            ip = self.mac_to_ip[mac]
+            ip = self.arp_mapping.get_ip(mac)
+            ip_query = self.arp_mapping.get_ip(mac_query)
             self.send_arp_reply(ip, vmac_new, ip_query, mac_query)
         del self.vmac_to_queries[vmac_old]
 
-        self.inception.delete_failover_log(i_conf.RPC_GATEWAY_FLOW)
+        if CONF.zookeeper_storage:
+            self.inception.delete_failover_log(i_conf.RPC_GATEWAY_FLOW)
